Use the gRPC dispatcher for event bus unary calls
Motivation:

- The unary path built `GrpcServerRequestImpl` and `GrpcServerResponseImpl` by hand instead of going through `GrpcDispatcher` like the streaming path, so the internal gRPC frame protocol was not the only way into a service method.
- `EventBusGrpcServerUnaryCall` and `EventBusGrpcClientUnaryCall` returned themselves from every handler setter and from `pause`/`resume`/`fetch`, duplicating handler fields already present in `EventBusGrpcStreamBase` while dropping the events and the back pressure on the floor.

Changes:

- Introduced `EventBusGrpcCallBase` holding the stream handlers, the inbound queue feeding them and the read stream flow control gating it, with `handleInboundMessage` and `handleInboundFlowing` hooks for the receive window accounting only streaming calls need.
- Refactored `EventBusGrpcStreamBase`, `EventBusGrpcServerUnaryCall` and `EventBusGrpcClientUnaryCall` to extend it, leaving `EventBusGrpcStreamBase` with the transport frames, the send window and the outbound queue.
- Updated `EventBusGrpcServerImpl.dispatchUnary` to wire a `GrpcDispatcher` the way `dispatchStreaming` does, and added `EventBusGrpcServerUnaryCall.init` publishing the call as a headers frame, a message frame and an end event.
- Dropped the `endHandler` registered by both dispatch methods, `GrpcDispatcher.handleHeadersFrame` overwrites it with its own before it can fire.
