+++ vertx-grpc-eventbus/src/main/java/io/vertx/grpc/eventbus/impl/EventBusGrpcStreamBase.java
@@ -3,15 +3,9 @@ package io.vertx.grpc.eventbus.impl;
 import com.google.protobuf.ByteString;
 import io.vertx.core.Closeable;
 import io.vertx.core.Future;
-import io.vertx.core.Handler;
 import io.vertx.core.Promise;
 import io.vertx.core.internal.ContextInternal;
 import io.vertx.grpc.common.GrpcMessage;
-import io.vertx.grpc.common.impl.GrpcFrame;
-import io.vertx.grpc.common.impl.GrpcFrameType;
-import io.vertx.grpc.common.impl.GrpcInboundStream;
-import io.vertx.grpc.common.impl.GrpcOutboundStream;
-import io.vertx.grpc.common.impl.GrpcStream;
 import io.vertx.grpc.eventbus.transport.v1alpha.Message;
 import io.vertx.grpc.eventbus.transport.v1alpha.TransportFrame;
 import io.vertx.grpc.eventbus.transport.v1alpha.WindowUpdate;
@@ -19,31 +13,20 @@ import io.vertx.grpc.eventbus.transport.v1alpha.WindowUpdate;
 import java.util.ArrayDeque;
 import java.util.Deque;
 
-abstract class EventBusGrpcStreamBase implements GrpcStream, Closeable {
+abstract class EventBusGrpcStreamBase extends EventBusGrpcCallBase implements Closeable {
 
   static final int DEFAULT_WINDOW = 64;
 
-  protected final ContextInternal context;
   protected final int window;
 
   private final Deque<MessageWrite> outboundQueue = new ArrayDeque<>();
 
-  private static final Object END_MARKER = new Object();
-  private final Deque<Object> inboundQueue = new ArrayDeque<>();
-  private boolean draining;
-
-  private Handler<GrpcFrame> frameHandler;
-  private Handler<Void> endHandler;
-  private Handler<Throwable> exceptionHandler;
-  private Handler<Void> drainHandler;
-
-  private boolean flowing = true;
   private int granted;
   private int sendWindow;
   private long sequence;
 
   EventBusGrpcStreamBase(ContextInternal context, int window) {
-    this.context = context;
+    super(context);
     this.window = window;
     this.granted = window;
   }
@@ -54,13 +37,19 @@ abstract class EventBusGrpcStreamBase implements GrpcStream, Closeable {
 
   abstract void handleRemoteEndpointDown(Throwable cause);
 
-  private void onInboundMessage() {
+  @Override
+  protected void handleInboundMessage() {
     granted--;
     topUpWindow();
   }
 
+  @Override
+  protected void handleInboundFlowing() {
+    topUpWindow();
+  }
+
   private void topUpWindow() {
-    if (flowing && granted <= window / 2) {
+    if (flowing() && granted <= window / 2) {
       int delta = window - granted;
       if (delta > 0) {
         granted += delta;
@@ -103,10 +92,7 @@ abstract class EventBusGrpcStreamBase implements GrpcStream, Closeable {
     sendWindow += delta;
     drainOutbound();
     if (wasFull && !writeQueueFull()) {
-      Handler<Void> h = drainHandler;
-      if (h != null) {
-        context.runOnContext(v -> h.handle(null));
-      }
+      handleDrain();
     }
   }
 
@@ -120,108 +106,6 @@ abstract class EventBusGrpcStreamBase implements GrpcStream, Closeable {
   @Override
   public boolean writeQueueFull() {
     return sendWindow <= 0 || !outboundQueue.isEmpty();
-  }
-
-  @Override
-  public GrpcOutboundStream setWriteQueueMaxSize(int maxSize) {
-    return this;
-  }
-
-  @Override
-  public GrpcOutboundStream drainHandler(Handler<Void> handler) {
-    this.drainHandler = handler;
-    return this;
-  }
-
-  @Override
-  public GrpcStream handler(Handler<GrpcFrame> handler) {
-    this.frameHandler = handler;
-    return this;
-  }
-
-  @Override
-  public GrpcStream endHandler(Handler<Void> handler) {
-    this.endHandler = handler;
-    return this;
-  }
-
-  @Override
-  public GrpcStream exceptionHandler(Handler<Throwable> handler) {
-    this.exceptionHandler = handler;
-    return this;
-  }
-
-  @Override
-  public GrpcInboundStream pause() {
-    flowing = false;
-    return this;
-  }
-
-  @Override
-  public GrpcInboundStream resume() {
-    return fetch(Long.MAX_VALUE);
-  }
-
-  @Override
-  public GrpcInboundStream fetch(long amount) {
-    if (amount > 0) {
-      flowing = true;
-      drainInbound();
-      topUpWindow();
-    }
-    return this;
-  }
-
-  protected void emit(GrpcFrame frame) {
-    inboundQueue.add(frame);
-    drainInbound();
-  }
-
-  protected void emitEnd() {
-    inboundQueue.add(END_MARKER);
-    drainInbound();
-  }
-
-  protected void emitException(Throwable t) {
-    inboundQueue.add(t);
-    drainInbound();
-  }
-
-  private void drainInbound() {
-    if (draining) {
-      return;
-    }
-    draining = true;
-    try {
-      while (flowing && !inboundQueue.isEmpty()) {
-        dispatchInbound(inboundQueue.poll());
-      }
-    } finally {
-      draining = false;
-    }
-  }
-
-  private void dispatchInbound(Object event) {
-    if (event == END_MARKER) {
-      Handler<Void> handler = endHandler;
-      if (handler != null) {
-        handler.handle(null);
-      }
-    } else if (event instanceof Throwable) {
-      Handler<Throwable> handler = exceptionHandler;
-      if (handler != null) {
-        handler.handle((Throwable) event);
-      }
-    } else {
-      GrpcFrame frame = (GrpcFrame) event;
-      Handler<GrpcFrame> handler = frameHandler;
-      if (handler != null) {
-        handler.handle(frame);
-      }
-      if (frame.type() == GrpcFrameType.MESSAGE) {
-        onInboundMessage();
-      }
-    }
   }
 
   private static final class MessageFrameWrite implements MessageWrite {
